feat(experiments): Pass in a reference to able to the verification experiemnts.
This allows each experiment to query able itself to determine whether it's
the active experiment, which results in *much* cleaner frontend code.

Now the frontend no longer needs to track which experiments are mutually
exclusive, it just runs through a list of possible experiments and sees
which are active.

Goes with https://github.com/mozilla/fxa-content-server/pull/4673